Gui: Move DlgAddPropertyVarSet -> DlgAddProperty
