+++ src/Gui/Dialogs/DlgExpressionInput.cpp
@@ -237,7 +237,7 @@ void DlgExpressionInput::initializeVarSets()
     comboBoxGroup.setObjectName(QStringLiteral("comboBoxGroup"));
     comboBoxGroup.setInsertPolicy(QComboBox::InsertAtTop);
     comboBoxGroup.setSizePolicy(QSizePolicy::Preferred, QSizePolicy::Preferred);
-    DlgAddPropertyVarSet::setWidgetForLabel("labelGroup", &comboBoxGroup, ui->formLayout);
+    DlgAddProperty::setWidgetForLabel("labelGroup", &comboBoxGroup, ui->formLayout);
     setTabOrder(ui->comboBoxVarSet, &comboBoxGroup);
     setTabOrder(&comboBoxGroup, ui->lineEditPropNew);
 
@@ -819,7 +819,7 @@ void DlgExpressionInput::onVarSetSelected(int /*index*/)
         return;
     }
 
-    DlgAddPropertyVarSet::populateGroup(comboBoxGroup, varSet);
+    DlgAddProperty::populateGroup(comboBoxGroup, varSet);
     preselectGroup();
     updateVarSetInfo();
     ui->lineEditPropNew->setFocus();
