+++ src/Gui/ViewProviderVarSet.cpp
@@ -44,7 +44,7 @@ ViewProviderVarSet::ViewProviderVarSet()
 bool ViewProviderVarSet::doubleClicked()
 {
     if (!dialog) {
-        dialog = std::make_unique<DlgAddPropertyVarSet>(getMainWindow(), this);
+        dialog = std::make_unique<DlgAddProperty>(getMainWindow(), this);
     }
 
     // Do not use exec() here because it blocks and prevents command Std_VarSet
